trunk: corrected integer kind

+++ src/base/dataio_pub.F90
@@ -43,7 +43,7 @@ module dataio_pub
    real, parameter :: piernik_hdf5_version = 1.17   !< output version
    real, parameter :: piernik_hdf5_version2 = 2.0   !< output version for multi-file, multi-domain I/O
    logical         :: use_v2_io                     !< prefer the new I/O format
-   integer         :: nproc_io                      !< how many processes do the I/O (v2 only)
+   integer(kind=4) :: nproc_io                      !< how many processes do the I/O (v2 only)
    logical         :: can_i_write                   !< .true. for processes allowed to write
 
    ! Buffer lengths used only in I/O routines
